[QR.1] DSP: MIRPipeline.elapsedSeconds → Double; LiveBeatDriftTracker.playbackTime widened (D-079)
`elapsedSeconds: Float` was incremented by `+= deltaTime` every frame.
After 30 minutes ULP ≈ 240 µs — smaller than the ±30 ms tight-match
window in `LiveBeatDriftTracker` but a guaranteed monotonic drift over
hours of listening.

`elapsedSeconds`, `lastOnsetRateTime`, and `lastRecordTime` promoted to
`Double`. `featureStability` cast to `Float` once at write site.
`LiveBeatDriftTracker.update(playbackTime:)` parameter widened to
`Double` so the precision survives onset matching and lock-state
computation. Test suite drift-tracker callers updated to pass `Double`
playbackTime directly (Swift widens `Float(t)` to `Double` automatically
via the parameter type, but the explicit cast was redundant).

Tests:
- New `elapsedSeconds_typeIsDouble` (compile-time assertion).
- New `elapsedSeconds_accumulatesAsDouble_isMoreAccurateThanFloat`
  (compares pipeline accumulator against parallel Float accumulator).

D-079, Audio+DSP D3.

+++ PhospheneEngine/Sources/DSP/LiveBeatDriftTracker.swift
@@ -215,9 +215,13 @@ public final class LiveBeatDriftTracker: @unchecked Sendable {
     /// `BeatDetector.Result.onsets[0]`, the playback time in seconds since
     /// track start (same domain as `BeatGrid.beats`), and the wall-clock
     /// `deltaTime` since the last call.
+    ///
+    /// `playbackTime` is `Double` (D-079, QR.1) so long-session callers
+    /// (`MIRPipeline.elapsedSeconds`) keep their full precision through the
+    /// onset-matching and lock-state path.
     public func update(
         subBassOnset: Bool,
-        playbackTime: Float,
+        playbackTime: Double,
         deltaTime: Float
     ) -> Result {
         lock.lock(); defer { lock.unlock() }
@@ -227,7 +231,7 @@ public final class LiveBeatDriftTracker: @unchecked Sendable {
             return Result(beatPhase01: 0, beatsUntilNext: 1, barPhase01: 0, beatsPerBar: 1, lockState: .unlocked)
         }
 
-        let pt = Double(playbackTime)
+        let pt = playbackTime
         let dt = Double(max(deltaTime, 0.001))
 
         if subBassOnset {
